fix(code-graph): restore CI module boundaries
